+++ src/lightkurve/targetpixelfile.py
@@ -540,17 +540,25 @@ class TargetPixelFile(object):
     def to_lightcurve(self, method="sap", corrector=None, **kwargs):
         """Performs photometry on the pixel data and returns a LightCurve object.
 
-        See the docstring of `aperture_photometry()` for valid
-        arguments if the method is 'aperture'.  Otherwise, see the docstring
-        of `prf_photometry()` for valid arguments if the method is 'prf'.
+        The valid keyword arguments depends on the method chosen:
+
+        - 'sap' or 'aperture': see the docstring of `extract_aperture_photometry()`
+        - 'prf': see the docstring of `extract_prf_photometry()`
+        - 'pld': see the docstring of `to_corrector()`
+
+        For methods 'sff' and 'cbv', they are syntactic shortcuts of:
+
+        - creating a lightcurve using 'sap' method,
+        - corrects the created lightcurve using `LightCurve.to_corrector()`
+          of the respective method.
 
         Parameters
         ----------
         method : 'aperture', 'prf', 'sap', 'sff', 'cbv', 'pld'.
-            Photometry method to use.
+            Photometry method to use. 'aperture' is an alias of 'sap'.
         **kwargs : dict
-            Extra arguments to be passed to the `aperture_photometry` or the
-            `prf_photometry` method of this class.
+            Extra arguments to be passed to the `extract_aperture_photometry()`, the
+            `extract_prf_photometry()`, or the `to_corrector()` method of this class.
 
         Returns
         -------
@@ -1360,21 +1368,21 @@ class TargetPixelFile(object):
         )
 
     def to_corrector(self, method="pld", **kwargs):
-        """Returns a `Corrector` instance to remove systematics.
+        """Returns a `~correctors.corrector.Corrector` instance to remove systematics.
 
         Parameters
         ----------
         methods : string
             Currently, only "pld" is supported.  This will return a
-            `PLDCorrector` class instance.
+            `~correctors.PLDCorrector` class instance.
         **kwargs : dict
             Extra keyword arguments to be passed on to the corrector class.
 
         Returns
         -------
-        correcter : `lightkurve.Correcter`
-            Instance of a Corrector class, which typically provides `correct()`
-            and `diagnose()` methods.
+        correcter : `~correctors.corrector.Corrector`
+            Instance of a Corrector class, which typically provides `~correctors.PLDCorrector.correct()`
+            and `~correctors.PLDCorrector.diagnose()` methods.
         """
         allowed_methods = ["pld"]
         if method == "sff":
@@ -2140,7 +2148,7 @@ class KeplerTargetPixelFile(TargetPixelFile):
             at each cadence. Defaults to 'sum'.
         centroid_method : str, 'moments' or 'quadratic'
             For the details on this arguments, please refer to the documentation
-            for `TargetPixelFile.estimate_centroids`.
+            for `estimate_centroids()`.
 
         Returns
         -------
@@ -2278,8 +2286,8 @@ class KeplerTargetPixelFile(TargetPixelFile):
             If `True`, fitting cadences will be distributed across multiple
             cores using Python's `multiprocessing` module.
         **kwargs : dict
-            Keywords to be passed to `tpf.get_model()` to create the
-            `TPFModel` object that will be fit.
+            Keywords to be passed to `get_model()` to create the
+            `~prf.TPFModel` object that will be fit.
 
         Returns
         -------
@@ -2799,7 +2807,7 @@ class TessTargetPixelFile(TargetPixelFile):
             at each cadence. Defaults to 'sum'.
         centroid_method : str, 'moments' or 'quadratic'
             For the details on this arguments, please refer to the documentation
-            for `TargetPixelFile.estimate_centroids`.
+            for `estimate_centroids()`.
 
         Returns
         -------